fix #93 separate test packages
